Move exceptions of uniquelist package into a subpackage
